Restore sanity.
I battled this for-loop to the brink of insanity, and emerged on the
other side victorious.  This monster behaves like a regular for-loop
except that it has the side "benefit" of crashing when the result set
has length 0 (if you take out the crash, it runs first with i=j=0 and
then i=j=-1 and then stops).